+++ src/tests/test_StudioClientConnection.cpp
@@ -82,7 +82,7 @@ Fixture::Fixture(int helloVersion, ConnectionTimings timings)
 
 void Fixture::connect()
 {
-  connection.connect("127.0.0.1", short(server.port()));
+  connection.connect("127.0.0.1", server.port());
 }
 
 bool Fixture::waitConnectedAndBootstrapped(int expectedBootstraps)
